# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DeliveryApp API",
        "description": "API for delivery management system",
        "contact": {
            "email": "admin@deliveryapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/root/logs": {
            "get": {
                "tags": [
                    "Root"
                ],
                "summary": "Get canceled requests log",
                "operationId": "3f1f1df983e01d9db9234387d35d1429",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Canceled requests log"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/root/stats": {
            "get": {
                "tags": [
                    "Root"
                ],
                "summary": "Get system statistics",
                "operationId": "a4227dc5fc191272401c92263b2edb28",
                "responses": {
                    "200": {
                        "description": "System statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "object"
                                        },
                                        "requests": {
                                            "type": "object"
                                        },
                                        "deliveries": {
                                            "type": "object"
                                        },
                                        "locations": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/root/users": {
            "get": {
                "tags": [
                    "Root"
                ],
                "summary": "List all users",
                "operationId": "65035070fbb00e592e0757b07bdf0f1a",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "root",
                                "cliente",
                                "entregador"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/root/status-history": {
            "get": {
                "tags": [
                    "Root"
                ],
                "summary": "Get full status history",
                "operationId": "3af18ed412c6204b566e0abef09a5c44",
                "parameters": [
                    {
                        "name": "entity_type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cliente",
                                "entregador",
                                "solicitacao",
                                "entrega"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status history"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}